refactor: group into CRMProvidersAccordion
